extract common conformance level logic to helper class
